requirements.txt
langchain==0.1.9
pyyaml==6.0.1
openai==1.13.3
tiktoken==0.6.0
faiss-cpu==1.8.0 # or faiss-gpu for gpu version
flask==3.0.2
ragas==0.1.3
datasets==2.18.0